--- src/main/java/com/example/g3bilabonnement/Controllers/PurchaseAgreementController.java
@@ -1,31 +1,0 @@
-package com.example.g3bilabonnement.Controllers;
-
-import com.example.g3bilabonnement.Service.PurchaseAgreementService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@RequestMapping("/purchase-agreement")
-@Controller
-public class PurchaseAgreementController {
-
-    @Autowired
-    PurchaseAgreementService purchaseAgreementService;
-
-    @GetMapping("/create")
-    public String showCreateForm() {
-        return "create";
-    }
-
-    @PostMapping("/create")
-    public String createPurchaseAgreement(@RequestParam("carId") int carId,
-                                          @RequestParam("pickuplocation") String pickupLocation) {
-
-        purchaseAgreementService.createPurchaseAgreement(carId, pickupLocation);
-
-        return "redirect:/success";
-    }
-}
