--- src/main/java/com/example/g3bilabonnement/Entities/Car.java
@@ -1,134 +1,0 @@
-package com.example.g3bilabonnement.Entities;
-
-public class Car {
-    private int id;
-    private String vehicleNumber;
-    private String vinNumber;
-    private String brand;
-    private String model;
-    private String equipmentLevel;
-    private String powerSourceType;
-    private String transmissionType;
-    private String status;
-    private double netPrice = 2000;
-    private double registrationTax;
-    private double co2Emissions;
-    private boolean prePurchased;
-    private DamageReport damageReport;
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public DamageReport getDamageReport() {
-        return damageReport;
-    }
-
-    public void setDamageReport(DamageReport damageReport) {
-        this.damageReport = damageReport;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getVehicleNumber() {
-        return vehicleNumber;
-    }
-
-    public void setVehicleNumber(String vehicleNumber) {
-        this.vehicleNumber = vehicleNumber;
-    }
-
-    public String getVinNumber() {
-        return vinNumber;
-    }
-
-    public void setVinNumber(String vinNumber) {
-        this.vinNumber = vinNumber;
-    }
-
-    public String getBrand() {
-        return brand;
-    }
-
-    public void setBrand(String brand) {
-        this.brand = brand;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public void setModel(String model) {
-        this.model = model;
-    }
-
-    public String getEquipmentLevel() {
-        return equipmentLevel;
-    }
-
-    public void setEquipmentLevel(String equipmentLevel) {
-        this.equipmentLevel = equipmentLevel;
-    }
-
-    public String getPowerSourceType() {
-        return powerSourceType;
-    }
-
-    public void setPowerSourceType(String powerSourceType) {
-        this.powerSourceType = powerSourceType;
-    }
-
-    public String getTransmissionType() {
-        return transmissionType;
-    }
-
-    public void setTransmissionType(String transmissionType) {
-        this.transmissionType = transmissionType;
-    }
-
-    public double getNetPrice() {
-        return netPrice;
-    }
-
-    public void setNetPrice(double netPrice) {
-        this.netPrice = netPrice;
-    }
-
-    public double getRegistrationTax() {
-        return registrationTax;
-    }
-
-    public void setRegistrationTax(double registrationTax) {
-        this.registrationTax = registrationTax;
-    }
-
-    public double getCo2Emissions() {
-        return co2Emissions;
-    }
-
-    public void setCo2Emissions(double co2Emissions) {
-        this.co2Emissions = co2Emissions;
-    }
-
-    public boolean isPrePurchased() {
-        return prePurchased;
-    }
-
-    public void setPrePurchased(boolean prePurchased) {
-        this.prePurchased = prePurchased;
-    }
-}
-
-
-
-
