--- src/main/java/com/example/g3bilabonnement/Entities/RentalAgreement.java
@@ -1,4 +1,0 @@
-package com.example.g3bilabonnement.Entities;
-
-public class RentalAgreement {
-}
