+++ src/main/java/com/example/g3bilabonnement/Service/PurchaseAgreementService.java
@@ -1,9 +1,8 @@
 package com.example.g3bilabonnement.Service;
 
-import com.example.g3bilabonnement.Entities.Car;
-import com.example.g3bilabonnement.Entities.FinalSettlement;
-import com.example.g3bilabonnement.Entities.PurchaseAgreement;
-import com.example.g3bilabonnement.Repository.CarRepository;
+import com.example.g3bilabonnement.entity.Car;
+import com.example.g3bilabonnement.entity.FinalSettlement;
+import com.example.g3bilabonnement.entity.PurchaseAgreement;
 import com.example.g3bilabonnement.Repository.PurchaseAgreementRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,21 +14,11 @@ public class PurchaseAgreementService {
     private PurchaseAgreementRepository purchaseAgreementRepository;
 
 
-    public void createPurchaseAgreement(int carId, String pickupLocation) {
+    public void add (PurchaseAgreement purchaseAgreement) {
 
 
-        Car car = new Car();
+        purchaseAgreementRepository.add(purchaseAgreement);
 
-        FinalSettlement finalSettlement = new FinalSettlement();
-
-        PurchaseAgreement purchaseAgreement = new PurchaseAgreement(carId, pickupLocation);
-        purchaseAgreement.setCar(car);
-        purchaseAgreement.setFinalSettlement(finalSettlement);
-        purchaseAgreement.setPickUpLocation(pickupLocation);
-
-
-        purchaseAgreementRepository.createPurchaseAgreement(purchaseAgreement);
-        //her ændres en bil status til sold
-        purchaseAgreementRepository.updateCarStatusToSold(carId, "sold");
+        purchaseAgreementRepository.updateCarStatusToSold(purchaseAgreement.getCar().getId(), "sold");
     }
 }
