+++ src/main/java/com/example/g3bilabonnement/entity/DamageSpecification.java
@@ -1,0 +1,4 @@
+package com.example.g3bilabonnement.entity;
+
+public class DamageSpecification {
+}
