+++ src/main/java/com/example/g3bilabonnement/repository/RentalAgreementRepository.java
@@ -18,29 +18,6 @@ public class RentalAgreementRepository {
     private JdbcTemplate jdbcTemplate;
     //TODO Teknisk set må Crud ikke eksistere i denne opgave
 
-//    private final RowMapper<RentalAgreement> rentalAgreementRowMapper = (rs, rowNum) -> {
-//        Car car = new Car();
-//        Renter renter = new Renter();
-//        Subscription subscription = new Subscription();
-//
-//        int id = rs.getInt("id");
-//        car.setId(rs.getInt("car_id"));
-//        renter.setId(rs.getInt("renter_id"));
-//        subscription.setId(rs.getInt("subscription_id"));
-//        LocalDate startDate = rs.getDate("start_date").toLocalDate();
-//        LocalDate endDate = rs.getDate("end_date").toLocalDate();
-//        double totalPrice = rs.getDouble("total_price");
-//
-//        return new RentalAgreement(id, car, subscription, renter, startDate, endDate, totalPrice);
-//    };
-
-//    public List<RentalAgreement> findAll() {
-//        return jdbcTemplate.query("SELECT * FROM RentalAgreement", rentalAgreementRowMapper);
-//    }
-//
-//    public RentalAgreement getById(int id) {
-//        return jdbcTemplate.queryForObject("SELECT * FROM RentalAgreement WHERE id = ?", rentalAgreementRowMapper, id);
-//    }
 
     public void add(RentalAgreement rentalAgreement) {
         String sql = "INSERT INTO rental_agreement (car_id, subscription_id, renter_id, start_date, end_date) " +
@@ -53,19 +30,4 @@ public class RentalAgreementRepository {
                 rentalAgreement.getEndDate());
     }
 
-//    public void update(RentalAgreement rentalAgreement) {
-//        String sql = "UPDATE RentalAgreement SET car_id = ?, subscription_id = ?, renter_id = ?, start_date = ?, end_date = ?" +
-//                "WHERE id = ?";
-//        jdbcTemplate.update(sql,
-//                rentalAgreement.getCar().getId(),
-//                rentalAgreement.getSubscription().getId(),
-//                rentalAgreement.getRenter().getId(),
-//                rentalAgreement.getStartDate(),
-//                rentalAgreement.getEndDate(),
-//                rentalAgreement.getId());
-//    }
-//
-//    public void delete(int id) {
-//        jdbcTemplate.update("DELETE FROM Artist WHERE ArtistID = ?", id);
-//    }
 }
