+++ src/main/resources/templates/createPurchaseAgreementResult.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Købsaftale Oprettet</title>
+</head>
+<body>
+<h1>Købsaftale Oprettet!</h1>
+<p>Bilen er blevet solgt, og købsaftalen er gemt i systemet.</p>
+<a href="/purchase-agreement/create">Opret en ny købsaftale</a>
+</body>
+</html>
