+++ src/main/resources/templates/purchaseAgreementMainPage.html
@@ -2,9 +2,12 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Velkommen</title>
 </head>
 <body>
-
+<h1>Velkommen til BilAbonnement.dk</h1>
+<p>Vælg en handling nedenfor:</p>
+<button onclick="location.href='/purchase-agreement/create'">Opret en købsaftale</button>
 </body>
 </html>
